Move Wasi bindings to separate file
